Reuse TurboSnap result types
